refactor(test_signals): restructure test_signals into several modules for splitting improvement
